[156] Move 'decodeAndCache' back to the utility class
