+++ app/src/main/java/net/aiscope/gdd_app/ui/metadata/SampleImagesAdapter.kt
@@ -96,7 +96,7 @@ private class ImageViewHolder(
         (itemView.tag as? Job)?.cancel()
         itemView.tag = uiScope.launch {
             (itemView as ImageView).setImageBitmap(null)
-            val bitmap = decodeSampledBitmapAndCache(
+            val bitmap = BitmapReader.decodeSampledBitmapAndCache(
                 image,
                 itemView.context.resources.getDimensionPixelSize(R.dimen.sample_image_thumbnail_width),
                 itemView.context.resources.getDimensionPixelSize(R.dimen.sample_image_thumbnail_height),
@@ -106,30 +106,3 @@ private class ImageViewHolder(
         }
     }
 }
-
-suspend fun decodeSampledBitmapAndCache(
-    image: File,
-    reqWidth: Int,
-    reqHeight: Int,
-    cacheDir: File
-): Bitmap = withContext(Dispatchers.IO) {
-    val cachedImage = File(
-        cacheDir,
-        "${image.nameWithoutExtension}_${reqWidth}x${reqHeight}.${image.extension}"
-    )
-    if (cachedImage.exists()) {
-        return@withContext BitmapFactory.decodeFile(cachedImage.absolutePath)
-    }
-    val bitmap = BitmapReader.decodeSampledBitmapFromResource(
-        image,
-        false,
-        MinimumSizeDownSampling(reqWidth, reqHeight)
-    )
-
-    //Write to cache for future access
-    bitmap.writeToFile(cachedImage, Bitmap.CompressFormat.JPEG)
-
-    bitmap
-}
-
-
